Fix uint8 overflows

+++ gym_super_mario_bros/smb_env.py
@@ -139,7 +139,7 @@ class SuperMarioBrosEnv(NESEnv):
     def _x_position(self):
         """Return the current horizontal position."""
         # add the current page 0x6d to the current x
-        return self.ram[0x6d] * 0x100 + self.ram[0x86]
+        return self.ram[0x6d].astype(int) * 0x100 + self.ram[0x86]
 
     @property
     def _left_x_position(self):
@@ -174,7 +174,7 @@ class SuperMarioBrosEnv(NESEnv):
         # check if Mario is above the viewport (the score board area)
         if self._y_viewport < 1:
             # y position overflows so we start from 255 and add the offset
-            return 255 + (255 - self._y_pixel)
+            return 255 + (255 - self._y_pixel.astype(int))
         # invert the y pixel into the distance from the bottom of the screen
         return 255 - self._y_pixel
